Fix for trac ticket #88 - Native 64bit MapGuide Support http://trac.osgeo.org/mapguide/ticket/88
Changes:
- Fix size_t portability issue - on 64bit systems this is 64 bits in size and not 32 bits.

Notes:
- This is only a partial fix as there are still more issues to resolve.


git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@3460 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Common/Foundation/System/AceStreamHelper.cpp
@@ -470,7 +470,7 @@ MgStreamHelper::MgStreamStatus MgAceStreamHelper::WriteBytes(const unsigned char
         if (bytesWritten > 0)
         {
             memmove(m_writeBuffer, &m_writeBuffer[bytesWritten], m_writeBufLength-bytesWritten);
-            m_writeBufLength -= (int) bytesWritten;
+            m_writeBufLength -= bytesWritten;
         }
 
         // and recurse to push more data in
@@ -506,7 +506,7 @@ MgStreamHelper::MgStreamStatus MgAceStreamHelper::Flush()
         if (bytesWritten > 0)
         {
             memmove(m_writeBuffer, &m_writeBuffer[bytesWritten], m_writeBufLength-bytesWritten);
-            m_writeBufLength -= (int) bytesWritten;
+            m_writeBufLength -= bytesWritten;
         }
     }
     return stat;
